Unify android theme with other targets and simplify layout
