+++ app/src/androidMain/kotlin/ir/ehsannarmani/compose_charts/MainActivity.kt
@@ -5,19 +5,15 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import ir.ehsannarmani.compose_charts.ui.App
-import ir.ehsannarmani.compose_charts.ui.theme.ComposeChartsTheme
 
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-
         enableEdgeToEdge()
         setContent {
-            ComposeChartsTheme(false) {
-                App()
-            }
+            App()
         }
     }
 }
