+++ app/src/commonMain/kotlin/ir/ehsannarmani/compose_charts/ui/AllSamples.kt
@@ -3,30 +3,32 @@ package ir.ehsannarmani.compose_charts.ui
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.statusBarsPadding
-import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.material3.LocalContentColor
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.SolidColor
-import androidx.compose.ui.graphics.drawscope.DrawStyle
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import ir.ehsannarmani.compose_charts.LineChart
 import ir.ehsannarmani.compose_charts.models.AnimationMode
 import ir.ehsannarmani.compose_charts.models.DotProperties
+import ir.ehsannarmani.compose_charts.models.LabelProperties
 import ir.ehsannarmani.compose_charts.models.Line
 import ir.ehsannarmani.compose_charts.models.PopupProperties
-import ir.ehsannarmani.compose_charts.models.StrokeStyle
 
 @Composable
-fun PhoneSample() {
-    LazyColumn(
-        modifier = Modifier
-            .statusBarsPadding(), verticalArrangement = Arrangement.spacedBy(28.dp),
-        contentPadding = PaddingValues(22.dp)
+fun TabletSample() {
+    LazyVerticalGrid(
+        columns = GridCells.Adaptive(250.dp),
+        verticalArrangement = Arrangement.spacedBy(28.dp),
+        horizontalArrangement = Arrangement.spacedBy(28.dp),
+        contentPadding = PaddingValues(28.dp)
     ) {
         item {
             ChartParent(Modifier) {
@@ -58,17 +60,20 @@ fun PhoneSample() {
                     animationMode = AnimationMode.None,
                     minValue = 0.0,
                     maxValue = 7.0,
-                    popupProperties = PopupProperties(
-                        textStyle = TextStyle.Default.copy(
-                            color = Color.White,
-                            fontSize = 12.sp
+                    labelProperties = labelProperties,
+                    labelHelperProperties = labelHelperProperties,
+                    indicatorProperties = indicatorProperties,
+                    popupProperties =
+                        PopupProperties(
+                            textStyle = TextStyle.Default.copy(
+                                color = Color.White,
+                                fontSize = 12.sp
+                            ),
+                            confirmDraw = {
+                                false
+                            }
                         ),
-                        confirmDraw = {
-                            false
-                        }
-                    ),
-
-                )
+                    )
             }
         }
         // Pie
@@ -93,7 +98,7 @@ fun PhoneSample() {
             ColumnSample3()
         }
 
-        // Row
+        //Row
         item {
             RowSample()
         }
@@ -132,5 +137,11 @@ fun PhoneSample() {
         item {
             LineSample9()
         }
+
     }
 }
+
+
+
+
+
