+++ app/src/commonMain/kotlin/ir/ehsannarmani/compose_charts/ui/Styles.kt
@@ -1,14 +1,23 @@
 package ir.ehsannarmani.compose_charts.ui
 
+import androidx.compose.material3.LocalContentColor
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.unit.sp
-import ir.ehsannarmani.compose_charts.models.LabelHelperProperties
-import ir.ehsannarmani.compose_charts.models.LabelProperties
+import ir.ehsannarmani.compose_charts.models.*
 
-val labelHelperProperties: LabelHelperProperties @Composable get()  = LabelHelperProperties(textStyle = TextStyle(fontSize = 12.sp, fontFamily = ubuntu, color = Color.White))
-val labelProperties: LabelProperties @Composable get()  = LabelProperties(
-    enabled = true,
-    textStyle = TextStyle(fontSize = 12.sp, fontFamily = ubuntu, color = Color.White)
-)
+val labelHelperProperties: LabelHelperProperties
+    @Composable get() =
+        LabelHelperProperties(textStyle = MaterialTheme.typography.labelMedium.copy(color = LocalContentColor.current))
+val labelProperties: LabelProperties
+    @Composable get() = LabelProperties(
+        enabled = true,
+        textStyle = MaterialTheme.typography.labelMedium.copy(color = LocalContentColor.current)
+    )
+
+val indicatorProperties: HorizontalIndicatorProperties
+    @Composable
+    get() = HorizontalIndicatorProperties(
+        textStyle = MaterialTheme.typography.labelMedium.copy(color = LocalContentColor.current),
+        count = IndicatorCount.CountBased(count = 4),
+        position = IndicatorPosition.Horizontal.Start,
+    )
